pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=42", "wheel"]

[tool.setuptools.packages.find]
exclude = [".vscode", "examples", "tests"]
include = ["aliyun*"]

[project]
authors = [{ name = "Lei Zhang", email = "jameszhang2880@gmail.com" }]
dependencies = [
    "python-dotenv",
    "alibabacloud-ecs20140526==7.0.0",
    "alibabacloud-cr20181201==2.0.1",
    "tenacity==9.1.2",
    "aiolimiter==1.2.1",
]
description = "aliyun open API"
license = { text = "MIT" }
name = "aliyun"
readme = "README.md"
requires-python = ">=3.10, <4.0"
version = "0.1.0"

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

# uv specific configuration
[tool.uv]
dev-dependencies = []
